refactor(add): delete the add-to-which-library picker
