+++ lib/mydia_web/components/library_components.ex
@@ -553,92 +553,4 @@ defmodule MydiaWeb.LibraryComponents do
     </button>
     """
   end
-
-  @doc """
-  The library chooser, rendered once per host LiveView.
-
-  An anchored dropdown inside the card cannot work here. The sidebar is
-  `z-40`, the mobile dock is `z-50` and the sticky header is `z-30`, so any
-  value a card claims either loses to the chrome or paints over it during
-  ordinary browsing, and a horizontal rail's `overflow` clips the menu
-  whatever its z-index. A page-level `.modal` is `position: fixed; inset: 0;
-  z-index: 999`, which sidesteps both problems.
-
-  The explicit `z-[1000]` puts the picker above the trending detail modal,
-  which is also a `.modal` at 999, without depending on which of the two
-  happens to come later in the DOM.
-
-  Libraries have no `name` column, only `path`, so each entry shows the
-  basename with the full path as secondary text.
-
-  `configure_event` is nil by default, which renders no Configure entry at
-  all: Dashboard and the media detail page's rail host this same dialog but
-  have no `open_add_config` handler, and a stray click there would crash the
-  LiveView with no matching `handle_event` clause. Discover is the only host
-  that passes it.
-  """
-  attr :picker, :map,
-    default: nil,
-    doc: "nil, or %{tmdb_id:, media_type:, title:, libraries:} for the card being added"
-
-  attr :event, :string, default: "add_to_library"
-  attr :on_cancel, :string, default: "close_library_picker"
-  attr :configure_event, :string, default: nil
-
-  def library_picker_dialog(assigns) do
-    ~H"""
-    <dialog
-      id="library-picker-dialog"
-      class="modal z-[1000]"
-      open={@picker != nil}
-      phx-window-keydown={@picker && @on_cancel}
-      phx-key="Escape"
-    >
-      <div :if={@picker} class="modal-box max-w-md">
-        <h3 class="font-bold text-lg">Add to which library?</h3>
-        <p class="text-sm text-base-content/60 mt-1 truncate">{@picker.title}</p>
-
-        <ul class="menu w-full mt-4 p-0">
-          <li :for={library <- @picker.libraries}>
-            <button
-              type="button"
-              data-test="library-picker-option"
-              phx-click={@event}
-              phx-value-library_path_id={library.id}
-              phx-value-tmdb_id={@picker.tmdb_id}
-              phx-value-media_type={@picker.media_type}
-              class="flex-col items-start gap-0"
-            >
-              <span class="font-medium">{Path.basename(library.path)}</span>
-              <span class="text-xs text-base-content/50 truncate w-full">{library.path}</span>
-            </button>
-          </li>
-          <li :if={@configure_event}>
-            <button
-              id="discover-configure-add"
-              type="button"
-              phx-click={@configure_event}
-              phx-value-tmdb_id={@picker.tmdb_id}
-              phx-value-media_type={@picker.media_type}
-            >
-              <.icon name="hero-adjustments-horizontal" class="w-4 h-4" /> Configure...
-            </button>
-          </li>
-        </ul>
-
-        <div class="modal-action">
-          <button type="button" class="btn btn-ghost" phx-click={@on_cancel}>Cancel</button>
-        </div>
-      </div>
-
-      <%!-- A <dialog> opened through the `open` attribute rather than
-           showModal() gets no ::backdrop and no Escape handling, which is why
-           the keydown above and this click target both exist. Same approach
-           as TrendingDetailModal. --%>
-      <div class="modal-backdrop bg-black/70">
-        <button type="button" phx-click={@on_cancel}>close</button>
-      </div>
-    </dialog>
-    """
-  end
 end
