+++ lib/mydia_web/live/dashboard_live/index.ex
@@ -47,7 +47,6 @@ defmodule MydiaWeb.DashboardLive.Index do
         |> assign(:detail_loading, false)
         |> assign(:movie_libraries, MediaAddHelpers.candidate_libraries(:movie))
         |> assign(:show_libraries, MediaAddHelpers.candidate_libraries(:tv_show))
-        |> assign(:library_picker, nil)
         |> assign(:add_config, nil)
         |> assign(:quality_profiles, Mydia.Settings.list_quality_profiles())
         |> load_dashboard_data()
@@ -75,7 +74,6 @@ defmodule MydiaWeb.DashboardLive.Index do
         |> assign(:detail_loading, false)
         |> assign(:movie_libraries, [])
         |> assign(:show_libraries, [])
-        |> assign(:library_picker, nil)
         |> assign(:add_config, nil)
         |> assign(:quality_profiles, Mydia.Settings.list_quality_profiles())
       end
@@ -160,14 +158,6 @@ defmodule MydiaWeb.DashboardLive.Index do
     end
   end
 
-  def handle_event("open_library_picker", params, socket) do
-    {:noreply, MediaAddHelpers.put_library_picker(socket, params)}
-  end
-
-  def handle_event("close_library_picker", _params, socket) do
-    {:noreply, MediaAddHelpers.clear_library_picker(socket)}
-  end
-
   def handle_event("open_add_config", params, socket) do
     {:noreply,
      MediaAddHelpers.put_add_config(
@@ -222,8 +212,6 @@ defmodule MydiaWeb.DashboardLive.Index do
         %{"tmdb_id" => provider_id, "media_type" => media_type} = params,
         socket
       ) do
-    socket = MediaAddHelpers.clear_library_picker(socket)
-
     case parse_event_media_type(media_type) do
       {:ok, media_type_atom} ->
         # An impatient double-click sends the event twice before the first
