+++ lib/mydia_web/live/discover_live/index.ex
@@ -73,7 +73,6 @@ defmodule MydiaWeb.DiscoverLive.Index do
       |> assign(:load_error, nil)
       |> assign(:detail_loading, false)
       |> assign(:libraries, [])
-      |> assign(:library_picker, nil)
       |> assign(:add_config, nil)
       |> assign(:quality_profiles, Settings.list_quality_profiles())
       |> GridDensity.assign_current()
@@ -163,8 +162,7 @@ defmodule MydiaWeb.DiscoverLive.Index do
        |> assign(:selected_year, nil)
        |> assign(:min_rating, nil)
        |> assign(:sort_by, "popularity.desc")
-       |> assign(:libraries, MediaAddHelpers.candidate_libraries(:movie))
-       |> assign(:library_picker, nil)}
+       |> assign(:libraries, MediaAddHelpers.candidate_libraries(:movie))}
     end
   end
 
@@ -225,14 +223,6 @@ defmodule MydiaWeb.DiscoverLive.Index do
     else
       {:noreply, socket}
     end
-  end
-
-  def handle_event("open_library_picker", params, socket) do
-    {:noreply, MediaAddHelpers.put_library_picker(socket, params)}
-  end
-
-  def handle_event("close_library_picker", _params, socket) do
-    {:noreply, MediaAddHelpers.clear_library_picker(socket)}
   end
 
   # Reached from the Configure caret. The preview is resolved from the current
@@ -297,8 +287,6 @@ defmodule MydiaWeb.DiscoverLive.Index do
         socket
       ) do
     with :ok <- Authorization.authorize_create_media(socket) do
-      socket = MediaAddHelpers.clear_library_picker(socket)
-
       case parse_event_media_type(media_type) do
         {:ok, media_type_atom} ->
           # An impatient double-click sends the event twice before the first
