+++ lib/mydia_web/live/media_live/show.ex
@@ -19,7 +19,6 @@ defmodule MydiaWeb.MediaLive.Show do
   alias MydiaWeb.MediaLive.Show.FranchiseEvents
   alias MydiaWeb.MediaLive.Show.RecommendationEvents
   alias MydiaWeb.MediaLive.Show.RecommendationComponents
-  alias MydiaWeb.MediaLive.Show.LibraryPickerEvents
   alias MydiaWeb.MediaLive.Show.AddConfigEvents
   alias MydiaWeb.Live.Helpers.RecommendationsExpanded
 
@@ -181,7 +180,6 @@ defmodule MydiaWeb.MediaLive.Show do
      |> assign(:adding_recommendation_tmdb_ids, MapSet.new())
      |> assign(:requesting_recommendation_id, nil)
      |> assign_new(:metadata_config, fn -> Mydia.Metadata.default_relay_config() end)
-     |> assign_new(:library_picker, fn -> nil end)
      |> assign(:add_config, nil)
      |> assign(
        :can_create_media,
@@ -563,15 +561,6 @@ defmodule MydiaWeb.MediaLive.Show do
   @impl true
   def handle_event("request_recommendation", params, socket),
     do: RecommendationEvents.request_recommendation(params, socket)
-
-  def handle_event("open_library_picker", params, socket),
-    do: LibraryPickerEvents.open_library_picker(params, socket)
-
-  def handle_event("close_library_picker", params, socket),
-    do: LibraryPickerEvents.close_library_picker(params, socket)
-
-  def handle_event("add_from_library_picker", params, socket),
-    do: LibraryPickerEvents.add_from_library_picker(params, socket)
 
   def handle_event("open_add_config", params, socket),
     do: AddConfigEvents.open_add_config(params, socket)
