+++ lib/mydia_web/live/media_live/show.html.heex
@@ -457,11 +457,6 @@
     />
   <% end %>
 
-  <MydiaWeb.LibraryComponents.library_picker_dialog
-    picker={@library_picker}
-    event="add_from_library_picker"
-  />
-
   <MydiaWeb.AddMediaComponents.add_config_modal
     config={@add_config}
     quality_profiles={@quality_profiles}
